requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib
bcrypt
pyjwt